+++ craft_parts/overlays/overlay_manager.py
@@ -62,6 +62,8 @@ class OverlayManager:
     :param part_list: A list of all parts in the project.
     :param base_layer_dir: The directory containing the overlay base, or None
         if the project doesn't use overlay parameters.
+    :param mount_package_sources: Configure chroot to use package sources from
+        the the host environment.
     :param cache_level: The number of part layers to be mounted before the
         package cache.
     """
@@ -73,12 +75,14 @@ class OverlayManager:
         part_list: list[Part],
         base_layer_dir: Path | None,
         cache_level: int,
+        mount_package_sources: bool = False,
     ) -> None:
         self._project_info = project_info
         self._part_list = part_list
         self._layer_dirs = [p.part_layer_dir for p in part_list]
         self._overlay_fs: OverlayFS | None = None
         self._base_layer_dir = base_layer_dir
+        self._mount_package_sources = mount_package_sources
         self._cache_level = cache_level
 
     @property
@@ -181,8 +185,11 @@ class OverlayManager:
         mount_dir = self._project_info.overlay_mount_dir
         # Ensure we always run refresh_packages_list by resetting the cache
         packages.Repository.refresh_packages_list.cache_clear()  # type: ignore[attr-defined]
+
         chroot.chroot(
-            mount_dir, _defer_evaluation(packages.Repository.refresh_packages_list)
+            mount_dir,
+            _defer_evaluation(packages.Repository.refresh_packages_list),
+            mount_package_sources=self._mount_package_sources,
         )
 
     def download_packages(self, package_names: list[str]) -> None:
@@ -197,7 +204,8 @@ class OverlayManager:
         chroot.chroot(
             mount_dir,
             _defer_evaluation(packages.Repository.download_packages),
-            package_names,
+            mount_package_sources=self._mount_package_sources,
+            args=(package_names,),
         )
 
     def install_packages(self, package_names: list[str]) -> None:
@@ -211,9 +219,10 @@ class OverlayManager:
         mount_dir = self._project_info.overlay_mount_dir
         chroot.chroot(
             mount_dir,
-            _defer_evaluation(packages.Repository.install_packages),
-            package_names,
-            refresh_package_cache=False,
+            packages.Repository.install_packages,
+            mount_package_sources=self._mount_package_sources,
+            args=(package_names,),
+            kwargs={"refresh_package_cache": False},
         )